剖析功能整理了，物件等以後有需要才閣做 fix #91
