+++ 臺灣言語工具/剖析/中研院/剖析用戶端.py
@@ -21,6 +21,7 @@ import time
 
 
 from 臺灣言語工具.斷詞.中研院.用戶端連線 import 用戶端連線
+from 臺灣言語工具.剖析.中研院.剖析結構化工具 import 剖析結構化工具
 
 class 剖析用戶端(用戶端連線):
 	def __init__(self, 主機='140.109.19.112', 連接埠=8000, 編碼='Big5',
@@ -30,10 +31,17 @@ class 剖析用戶端(用戶端連線):
 		self.連接埠 = 連接埠
 		self.帳號 = 帳號
 		self.密碼 = 密碼
-	def 剖析物件(self):
-		pass
-	def 語句剖析後結構化(self):
-		pass
+		
+		self.結構化工具 = 剖析結構化工具()
+	def 語句剖析後結構化(self, 語句, 等待=5, 一定愛成功=False):
+		語句結果=self.語句剖析做語句(語句, 等待, 一定愛成功)
+		結構化結果=[]
+		for 一逝字 in 語句結果:
+			一逝結構化=[]
+			for 一句 in 一逝字:
+				一逝結構化.append(self.結構化工具.結構化剖析結果(一句))
+			結構化結果.append(一逝結構化)
+		return 結構化結果
 	def 語句剖析做語句(self, 語句, 等待=5, 一定愛成功=False):
 		# 官方功能無記錄原本換逝資訊，所以愛一逐一擺
 		結果 = []
@@ -54,5 +62,4 @@ class 剖析用戶端(用戶端連線):
 						raise
 				else:
 					break
-		print('剖析', 結果)
 		return 結果
