+++ 試驗/剖析/中研院剖析用戶端試驗.py
@@ -17,66 +17,43 @@
 感謝您的使用與推廣～～勞力！承蒙！
 """
 import unittest
+from unittest.mock import patch
 
 
 from 臺灣言語工具.剖析.中研院.剖析用戶端 import 剖析用戶端
 
+@patch('臺灣言語工具.剖析.中研院.剖析用戶端.剖析用戶端.語句剖析做語句')
 class 中研院剖析用戶端試驗(unittest.TestCase):
 	def setUp(self):
 		self.用戶端 = 剖析用戶端()
 	def tearDown(self):
 		pass
 
-	@unittest.expectedFailure
-	def test_物件剖一句(self):
-		self.assertEqual(self.用戶端.剖析('我想吃飯'), [[
+	def test_結構剖一句(self, 語句剖析做語句mock):
+		語句剖析做語句mock.return_value=[[
  			'#1:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vi:吃飯))#',
- 			]])
-	@unittest.expectedFailure
-	def test_物件剖一逝字(self):
-		self.assertEqual(self.用戶端.剖析('我想吃飯。我想吃很多飯。'), [[
- 			'#1:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vi:吃飯))#。(PERIODCATEGORY)',
- 			'#2:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vt:吃|NP(DET:很多|Head:N:飯)))#。(PERIODCATEGORY)'],
- 			])
-	@unittest.expectedFailure
-	def test_物件剖兩逝字(self):
-		self.assertEqual(self.用戶端.剖析('我想吃飯。我想吃很多飯。\n我吃飽了。'), [[
-	 			'#1:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vi:吃飯))#。(PERIODCATEGORY)',
-	 			'#2:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vt:吃|NP(DET:很多|Head:N:飯)))#。(PERIODCATEGORY)',
- 			], [
-	 			'#1:1.[0] S(NP(Head:N:我)|Head:Vi:吃飽|T:了)#。(PERIODCATEGORY)',
- 			],
-			])
-	@unittest.expectedFailure
-	def test_物件剖濟逝字(self):
-		self.assertEqual(self.用戶端.剖析('\n\n我想吃飯。我想吃很多飯。\n\n  \n\n  　 \n\n我吃飽了。\n\n'), [[
-	 			'#1:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vi:吃飯))#。(PERIODCATEGORY)',
-	 			'#2:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vt:吃|NP(DET:很多|Head:N:飯)))#。(PERIODCATEGORY)',
- 			], [
-	 			'#1:1.[0] S(NP(Head:N:我)|Head:Vi:吃飽|T:了)#。(PERIODCATEGORY)',
- 			],
-			])
-	@unittest.expectedFailure
-	def test_物件剖著大於符號(self):
-		self.assertEqual(self.用戶端.剖析('我想) :>'), [[
-			'#1:1.[0] %(NP(Head:N:我)|Vt:想|COLONCATEGORY::)#&gt;(PARENTHESISCATEGORY)',
-			]])
-	@unittest.expectedFailure
-	def test_物件剖著小於符號是空的(self):
-		self.assertEqual(self.用戶端.剖析('我想) :<'), [[
-			]])
-
-	def test_結構剖一句(self):
+ 			]]
 		self.assertEqual(self.用戶端.語句剖析後結構化('我想吃飯'), [[
  			('1', ['S', ['NP', ('我', 'N', 'Head')], ('想', 'Vt', 'Head'), ['VP', ('吃飯', 'Vi', 'Head')]], ''),
  			]])
-	def test_結構剖一逝字(self):
+	def test_結構剖一逝字(self, 語句剖析做語句mock):
+		語句剖析做語句mock.return_value=[[
+		 			'#1:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vi:吃飯))#。(PERIODCATEGORY)',
+ 			'#2:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vt:吃|NP(DET:很多|Head:N:飯)))#。(PERIODCATEGORY)',
+ 			]]
 		self.assertEqual(self.用戶端.語句剖析後結構化('我想吃飯。我想吃很多飯。'), [[
  			('1', ['S', ['NP', ('我', 'N', 'Head')], ('想', 'Vt', 'Head'), ['VP', ('吃飯', 'Vi', 'Head')]], '。(PERIODCATEGORY)'),
 		 	('2', ['S', ['NP', ('我', 'N', 'Head')], ('想', 'Vt', 'Head'), ['VP', ('吃', 'Vt', 'Head'), ['NP', ('很多', 'DET'), ('飯', 'N', 'Head')]]], '。(PERIODCATEGORY)'),
  			],
  			])
-	def test_結構剖兩逝字(self):
+	def test_結構剖有換逝句(self, 語句剖析做語句mock):
+		語句剖析做語句mock.return_value= [[
+		 			'#1:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vi:吃飯))#。(PERIODCATEGORY)',
+		 			'#2:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vt:吃|NP(DET:很多|Head:N:飯)))#。(PERIODCATEGORY)',
+	 			], [
+		 			'#1:1.[0] S(NP(Head:N:我)|Head:Vi:吃飽|T:了)#。(PERIODCATEGORY)',
+	 			],
+			]
 		self.assertEqual(self.用戶端.語句剖析後結構化('我想吃飯。我想吃很多飯。\n我吃飽了。'), [[
 	 			('1', ['S', ['NP', ('我', 'N', 'Head')], ('想', 'Vt', 'Head'), ['VP', ('吃飯', 'Vi', 'Head')]], '。(PERIODCATEGORY)'),
 		 		('2', ['S', ['NP', ('我', 'N', 'Head')], ('想', 'Vt', 'Head'), ['VP', ('吃', 'Vt', 'Head'), ['NP', ('很多', 'DET'), ('飯', 'N', 'Head')]]], '。(PERIODCATEGORY)'),
@@ -84,51 +61,14 @@ class 中研院剖析用戶端試驗(unittest.TestCase):
 				('1', ['S', ['NP', ('我', 'N', 'Head')], ('吃飽', 'Vi', 'Head'), ('了', 'T')], '。(PERIODCATEGORY)'),
  			],
 			])
-	def test_結構剖濟逝字(self):
-		self.assertEqual(self.用戶端.語句剖析後結構化('\n\n我想吃飯。我想吃很多飯。\n\n  \n\n  　 \n\n我吃飽了。\n\n'), [[
-	 			('1', ['S', ['NP', ('我', 'N', 'Head')], ('想', 'Vt', 'Head'), ['VP', ('吃飯', 'Vi', 'Head')]], '。(PERIODCATEGORY)'),
-		 	('2', ['S', ['NP', ('我', 'N', 'Head')], ('想', 'Vt', 'Head'), ['VP', ('吃', 'Vt', 'Head'), ['NP', ('很多', 'DET'), ('飯', 'N', 'Head')]]], '。(PERIODCATEGORY)'),
- 			], [
-				('1', ['S', ['NP', ('我', 'N', 'Head')], ('吃飽', 'Vi', 'Head'), ('了', 'T')], '。(PERIODCATEGORY)'),
- 			],
-			])
-	def test_結構剖著大於符號(self):
+	def test_結構剖著大於符號(self, 語句剖析做語句mock):
+		語句剖析做語句mock.return_value= [[
+			'#1:1.[0] %(NP(Head:N:我)|Vt:想|COLONCATEGORY::)#&gt;(PARENTHESISCATEGORY)',
+			]]
 		self.assertEqual(self.用戶端.語句剖析後結構化('我想) :>'), [[
 			('1', ['%', ['NP', ('我', 'N', 'Head')], ('想', 'Vt'), ('', '', 'COLONCATEGORY')], '&gt;(PARENTHESISCATEGORY)'),
 			]])
-	def test_結構剖著小於符號是空的(self):
+	def test_結構剖著小於符號是空的(self, 語句剖析做語句mock):
+		語句剖析做語句mock.return_value= [[]]
 		self.assertEqual(self.用戶端.語句剖析後結構化('我想) :<'), [[
-			]])
-
-	def test_語句剖一句(self):
-		self.assertEqual(self.用戶端.語句剖析做語句('我想吃飯'), [[
- 			'#1:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vi:吃飯))#',
- 			]])
-	def test_語句剖一逝字(self):
-		self.assertEqual(self.用戶端.語句剖析做語句('我想吃飯。我想吃很多飯。'), [[
- 			'#1:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vi:吃飯))#。(PERIODCATEGORY)',
- 			'#2:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vt:吃|NP(DET:很多|Head:N:飯)))#。(PERIODCATEGORY)'],
- 			])
-	def test_語句剖兩逝字(self):
-		self.assertEqual(self.用戶端.語句剖析做語句('我想吃飯。我想吃很多飯。\n我吃飽了。'), [[
-	 			'#1:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vi:吃飯))#。(PERIODCATEGORY)',
-	 			'#2:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vt:吃|NP(DET:很多|Head:N:飯)))#。(PERIODCATEGORY)',
- 			], [
-	 			'#1:1.[0] S(NP(Head:N:我)|Head:Vi:吃飽|T:了)#。(PERIODCATEGORY)',
- 			],
-			])
-	def test_語句剖濟逝字(self):
-		self.assertEqual(self.用戶端.語句剖析做語句('\n\n我想吃飯。我想吃很多飯。\n\n  \n\n  　 \n\n我吃飽了。\n\n'), [[
-	 			'#1:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vi:吃飯))#。(PERIODCATEGORY)',
-	 			'#2:1.[0] S(NP(Head:N:我)|Head:Vt:想|VP(Head:Vt:吃|NP(DET:很多|Head:N:飯)))#。(PERIODCATEGORY)',
- 			], [
-	 			'#1:1.[0] S(NP(Head:N:我)|Head:Vi:吃飽|T:了)#。(PERIODCATEGORY)',
- 			],
-			])
-	def test_語句剖著大於符號(self):
-		self.assertEqual(self.用戶端.語句剖析做語句('我想) :>'), [[
-			'#1:1.[0] %(NP(Head:N:我)|Vt:想|COLONCATEGORY::)#&gt;(PARENTHESISCATEGORY)',
-			]])
-	def test_語句剖著小於符號是空的(self):
-		self.assertEqual(self.用戶端.語句剖析做語句('我想) :<'), [[
 			]])
