+++ 試驗/走全部試驗.py
@@ -150,10 +150,10 @@ if __name__ == '__main__':
 		連詞揀集內組試驗,
 # 	辭典連詞斷詞試驗,
 		中研院斷詞用戶端試驗,
+		中研院剖析用戶端試驗,
 		])
 	if 整合試驗:
 		愛走的試驗.extend([
-				中研院剖析用戶端試驗,
 				中研院斷詞用戶端整合試驗,
 				中研院剖析用戶端整合試驗,
 			])
